# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Kotemon Jastip API
  description: |
    API for Kotemon Jastip - A personal shopping service from Japan to Indonesia.

    ## Authentication

    This API uses two authentication methods:

    1. **Cookie-based** (`cookieAuth`): For regular user sessions after Google OAuth
       - Session cookie name: `jastip_session`
       - Used by: `/api/auth/*` endpoints (except login/init)

    2. **Bearer Token** (`bearerAuth`): For admin access
       - Header: `Authorization: Bearer <admin_password>`
       - Used by: All `/api/admin/*` endpoints

    ## Pricing Model

    The API uses a three-field pricing model:

    - **basePriceYen** (integer): Original price in Japanese Yen (admin enters)
    - **basePriceRp** (integer): Calculated base price in IDR = `basePriceYen × exchangeRate`
    - **sellingPriceRp** (integer): Final selling price in IDR (manual entry)

    **Profit calculation:** `profit = sellingPriceRp - basePriceRp`

    The `basePriceRp` is automatically calculated when creating or updating items based on the current
    `exchangeRate` stored in settings. When `basePriceYen` is updated via PATCH, `basePriceRp` is
    recalculated using the current exchange rate.

    ## Environment Variables

    Required environment variables:
    - `JWT_SECRET` - Secret for signing session tokens
    - `GOOGLE_CLIENT_SECRET` - Google OAuth client secret
    - `DB` - D1 database binding
    - `PHOTOS_BUCKET` - R2 bucket for photo storage
...
  contact:
...
servers:
...
paths:
...
  /auth/google:
    get:
      summary: Initiate Google OAuth flow
      description: |
        Redirects the user to Google's OAuth consent screen.
        After authorization, user is redirected to `/api/auth/google/callback`.
      tags:
        - Authentication
      operationId: initiateGoogleOAuth
      parameters:
        - name: frontend
          in: query
          description: Frontend URL to redirect back to after authentication
          schema:
            type: string
            format: uri
            default: "https://jastip.dkotama.com"
            example: "https://jastip.dkotama.com"
      responses:
        '302':
          description: Redirects to Google OAuth consent screen
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: Google OAuth authorization URL
            Set-Cookie:
              schema:
                type: string
              description: OAuth state cookie for CSRF protection
  /auth/google/callback:
    get:
      summary: Google OAuth callback
      description: |
        Handles the callback from Google after OAuth authorization.
        For existing users: Sets session cookie and redirects to frontend with token.
        For new users: Creates temp session and redirects to token verification page.
      tags:
        - Authentication
      operationId: googleOAuthCallback
      parameters:
        - name: code
          in: query
          description: Authorization code from Google
          schema:
            type: string
        - name: state
          in: query
          description: State parameter for CSRF protection
          schema:
            type: string
        - name: error
          in: query
          description: Error message if OAuth was denied
          schema:
            type: string
      responses:
        '302':
          description: Redirects to frontend with token or error
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: |
                Redirect URL:
                - Existing users: `{frontend}?token={jwt}`
                - New users: `{frontend}/verify-token?tempToken={jwt}`
                - Errors: `{frontend}?error={error_code}`
            Set-Cookie:
              schema:
                type: string
              description: Session or temp session cookie
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/verify-token:
    post:
      summary: Verify invite token and create user
      description: |
        Verifies an invite token and completes user registration for new OAuth users.
        Requires a valid temp session (from cookie or request body).
      tags:
        - Authentication
      operationId: verifyInviteToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: Invite token code (e.g., "KOTEMON2025")
                  example: "KOTEMON2025"
                tempToken:
                  type: string
                  description: Temp session token (for cross-origin requests)
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Token verified and user created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: |
                Two cookies are set:
                1. Clears the temp session cookie
                2. Sets the session cookie (`jastip_session`)
        '400':
          description: Invalid or expired invite token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Session expired or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      summary: Logout user
      description: Clears the session cookie and logs out the user
      tags:
        - Authentication
      operationId: logout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears the session cookie
  /auth/me:
    get:
      summary: Get current user info
      description: Retrieves information about the currently authenticated user
      tags:
        - Authentication
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User info retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: User account has been revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/test/mock-oauth:
    post:
      summary: Mock OAuth for testing (DEVELOPMENT ONLY)
      description: |
        Creates a mock Google OAuth user for local testing.
        Only available on localhost.
      tags:
        - Authentication
      operationId: mockOAuth
      responses:
        '200':
          description: Mock user created or logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: "New user - needs token verification"
                      tempToken:
                        type: string
                        description: Temp token for new users
                      sessionToken:
                        type: string
                        description: Session token for existing users
                      user:
                        oneOf:
                          - $ref: '#/components/schemas/User'
                          - $ref: '#/components/schemas/TempOAuthSession'
        '403':
          description: Not available on non-localhost environments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Admin Routes
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jastip_session
      description: Session cookie for authenticated users (7-day expiry)
    bearerAuth:
      type: http
      scheme: bearer
      description: Admin password used as bearer token
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        photoUrl:
          type: string
          format: uri
          nullable: true
          example: "https://lh3.googleusercontent.com/abc123"
    TempOAuthSession:
      type: object
      properties:
        googleId:
          type: string
          example: "123456789012345678901"
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "John Doe"
        photoUrl:
          type: string
          format: uri
          nullable: true
          example: "https://example.com/avatar.jpg"

    # Admin schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid request"
  responses:
...
    Unauthorized:
      description: Unauthorized - missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Not authenticated"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"